requirements.txt
certifi==2026.1.4
GeoAlchemy2==0.18.1
geopandas==1.1.2
greenlet==3.3.0
numpy==2.4.0
packaging==25.0
pandas==2.3.3
psycopg2-binary==2.9.11
pyogrio==0.12.1
pyproj==3.7.2
python-dateutil==2.9.0.post0
pytz==2025.2
shapely==2.1.2
six==1.17.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
tzdata==2025.3
Flask==3.0.0
flask-cors==4.0.0
